새로운 DB에 맞춰서 새롭게

+++ fashion-api/src/main/java/com/kdt03/fashion_api/domain/dto/ProductMapColumnDTO.java
@@ -10,6 +10,6 @@ public class ProductMapColumnDTO {
     private List<String> productIds;
     private List<String> productNames;
     private List<String> styles;
-    private List<Float> xCoords;
-    private List<Float> yCoords;
+    private List<Double> xCoords;
+    private List<Double> yCoords;
 }
